PythonBackend_ComplianceCommsAgent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.6.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
azure-identity>=1.15.0
PyPDF2==3.0.1
aiohttp>=3.8.0
